backend/requirements.txt
python-dotenv
langchain
fastapi
tiktoken
uvicorn
openai